[init-koin] koin 초기화가 1회만 이뤄질 수 있도록, composeApp 모듈의 웹 대응을 feature:main 모듈로 이관
