+++ feature/main/src/commonMain/composeResources/drawable/ic_battery.xml
