+++ feature/main/src/commonMain/composeResources/drawable/ic_signal.xml
