+++ feature/main/src/commonMain/composeResources/drawable/ic_wifi.xml
