+++ feature/main/src/commonMain/composeResources/drawable/img_front_camera.png
