+++ feature/main/src/nonWasmJsMain/kotlin/com/droidknights/app/feature/main/MainScreen.nonWasmJs.kt
@@ -1,0 +1,9 @@
+package com.droidknights.app.feature.main
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+actual fun MainScreen(modifier: Modifier) {
+    MainContent(modifier = modifier)
+}
